Safari-611.1.10/LayoutTests/crypto/subtle/ec-import-pkcs8-key-export-jwk-key-expected.txt

Test importing a P-384 PKCS8 ECDH key and exporting it in JWK format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS privateKey.kty is expectedJwkKey.kty
PASS privateKey.crv is expectedJwkKey.crv
PASS privateKey.x is expectedJwkKey.x
PASS privateKey.y is expectedJwkKey.y
PASS privateKey.d is expectedJwkKey.d
PASS privateKey.key_ops is expectedJwkKey.key_ops
PASS privateKey.ext is expectedJwkKey.ext
PASS successfullyParsed is true

TEST COMPLETE
